refactor(ast_tools): clarify code (#12919)
We can now use `if let` chains. Previously we had to use an early return pattern in some places because one branch had multiple nested `if` statements. Now that they're combined into a single `if`, we don't need to use this pattern any more, and can convert to `if` / `else`. This isn't always shorter in all cases, but in my opinion is a bit clearer.