docs(tasks): correct comment to match code (#12923)
Happened to come across this. The comment didn't match the code here.

I *assume* the code is right and the comment is outdated, so this PR
amends the comment. But I don't know what this code does, so maybe it's
the other way around and the code should be changed to match the
comment?