perf(lexer): only check for hashbang at start of file (#12521)
Small optimization to lexer. A hashbang can only appear at very start of file, so only check for hashbang when getting first token. This streamlines the byte handler for `#`, because a `#` anywhere else can only be a private identifier.

Note: `self.token.set_is_on_new_line(true);` in `read_hashbang_comment` is not required, because it's always `true` already.